third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/referer.https-expected.txt

CONSOLE ERROR: line 24: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1:8444') does not match the recipient window's origin ('https://web-platform.test:8444').
This is a testharness.js-based test.
Harness Error. harness_status.status = 2 , harness_status.message = null
TIMEOUT Verify the referer Test timed out
Harness: the test ran to completion.
